Base: simplify UnitsSchemas management
Fixes: Maintaining schemas is difficult and error-prone

- Facilitate easy schemas add, remove, change, etc.
- Remove 14 files containing approx 2,190 lines of if/else code and data
- Place data in one file (UnitsSchemasData.h) using a normalized structure (including special functions)
- Isolate and simplify data operations (code)
- Remove schemas enum to keep data independent of code
- Separate responsibilities: Specifications, data, schemas, schema
- Add schema data 'isDefault'
- Add schema data name
- Prefer algorithms to raw loops
- Add schemas unit tests
- Tweak quantity unit tests

+++ src/Gui/Selection/Selection.cpp
@@ -841,7 +841,7 @@ QString getPreselectionInfo(const char* documentName,
 {
     auto pts = schemaTranslatePoint(x, y, z, precision);
 
-    int numberDecimals = std::min(6, Base::UnitsApi::getDecimals());
+    int numberDecimals = std::min(6, static_cast<int>(Base::UnitsApi::getDecimals()));
 
     QString message = QStringLiteral("Preselected: %1.%2.%3 (%4 %5, %6 %7, %8 %9)")
         .arg(QString::fromUtf8(documentName))
